Basic breakup into modules with minimal changes.
