+++ .gitignore
@@ -1,2 +1,3 @@
 coverage
 node_modules
+dist
