+++ bin/test.js
@@ -1,0 +1,23 @@
+#!/usr/bin/env node
+require('qunitjs');
+
+if (process.argv[2] === 'testem') {
+  // use basic TAP output for testem
+  var sys = require('sys');
+  var num = 1;
+  QUnit.begin(function (details) {
+    sys.puts('1..'+details.totalTests);
+  });
+  QUnit.testDone(function (details) {
+    sys.puts((details.failed ? 'not ok ' : 'ok ') + (num++) + ' - ' + details.module + ' - ' + details.name);
+  });
+} else {
+  // otherwise use a nicer test reporter
+  var qe = require('qunit-extras');
+  qe.runInContext(global);
+}
+
+require('../tests/index.js');
+
+// run tests
+QUnit.load();
