+++ lib/index.js
@@ -1,0 +1,3 @@
+export { default as Bin } from './bin';
+export { default as ShelfFirst } from './shelf-first';
+export { default as FixedGrid } from './fixed-grid';
