+++ lib/shelf-first.js
@@ -1,98 +1,10 @@
-// I don't think this should have args
-function Bin(content, width) {
-  this.width = width || 0;
-  this.content = content;
-}
+import {
+  default as Bin,
+  rangeError
+} from './bin';
+import Entry from './entry';
 
-function mustImplement(name) {
-  return function() {
-    throw new TypeError("MustImplement: " + name );
-  };
-}
-
-function Position(x, y) {
-  this.x = x;
-  this.y = y;
-}
-
-// @param height: "known to have atleast this might height available"
-function Entry(height, width, x, y) {
-  this.height   = height;
-  this.width    = width;
-  this.position = new Position(x, y);
-}
-
-// abstract
-Bin.prototype.objectAt = function (collection, index) {
-  return collection[index];
-};
-
-// abstract: return coordinates of element at index.
-//
-// @param index: index of the element in content
-// @param width: viewport width.
-// @returns {x, y} coordinates of element at index.
-//
-// May reset cached viewport width.
-Bin.prototype.position = mustImplement('position');
-
-// abstract: reset internal state to be anchored at index.
-// @param index: index of the element in content
-Bin.prototype.flush = mustImplement('flush');
-
-// abstract: return total content height given viewport width.
-// @param width: viewport width
-//
-// May reset cached viewport width.
-Bin.prototype.height = mustImplement('height');
-
-// abstract: true if layout places more than one object on a line.
-Bin.prototype.isGrid = mustImplement('isGrid');
-
-function rangeError(length, index) {
-  throw new RangeError("Parameter must be within: [" + 0 + " and " + length + ") but was: " + index);
-}
-
-function insufficientArguments(actual, expected) {
-  throw new TypeError("Insufficent Arguments expected: " + expected + " but got " + actual + "");
-}
-
-// abstract: returns number of elements in content.
-Bin.prototype.length = function () {
-  return this.content.length;
-};
-
-// maximum offset of content wrt to viewport
-// The amount by which content (after being layed out) is taller than
-// the viewport.
-Bin.prototype.maxContentOffset = function Bin_maxContentOffset(width, height) {
-  var contentHeight = this.height(width);
-  var maxOffset = Math.max(contentHeight - height, 0);
-  return maxOffset;
-}
-
-// abstract: returns index of first visible item.
-// @param topOffset: scroll position
-// @param width: width of viewport
-// @param height: height of viewport
-//
-Bin.prototype.visibleStartingIndex = mustImplement('visibleStartingIndex');
-
-// abstract: returns number of items visible in viewport.
-// @param topOffset: scroll position
-// @param width: width of viewport
-// @param height: height of viewport
-Bin.prototype.numberVisibleWithin = mustImplement('numberVisibleWithin');
-
-Bin.prototype.heightAtIndex = function (index) {
-  return this.content[index].height;
-};
-
-Bin.prototype.widthAtIndex = function (index) {
-  return this.content[index].width;
-};
-
-function ShelfFirst(content, width) {
+export default function ShelfFirst(content, width) {
   this._super$constructor(content, width);
   this._positionEntries = [];
 }
@@ -357,102 +269,3 @@ ShelfFirst.prototype.visibleStartingIndex = function (topOffset, width, visibleH
 
   return -1;
 };
-
-function FixedGrid(content, elementWidth, elementHeight) {
-  this._elementWidth =  elementWidth;
-  this._elementHeight =  elementHeight;
-
-  this._super$constructor(content);
-}
-
-FixedGrid.prototype = Object.create(Bin.prototype);
-FixedGrid.prototype._super$constructor = Bin;
-
-FixedGrid.prototype.flush = function (/* index, to */) {
-
-};
-
-FixedGrid.prototype.isGrid = function (width) {
-  return (Math.floor(width / this.widthAtIndex(0)) || 1) > 1;
-};
-
-FixedGrid.prototype.visibleStartingIndex = function (topOffset, width, height) {
-  topOffset = Math.min(topOffset, this.maxContentOffset(width, height));
-  var columns = Math.floor(width / this.widthAtIndex(0)) || 1;
-
-  return Math.floor(topOffset / this.heightAtIndex(0)) * columns;
-};
-
-FixedGrid.prototype.numberVisibleWithin = function (topOffset, width, height, withPadding) {
-  var startingIndex = this.visibleStartingIndex(topOffset, width, height);
-  var columns = Math.floor(width / this.widthAtIndex(0)) || 1;
-  var length = this.length();
-
-  var rowHeight = this.heightAtIndex(0);
-  var rows = Math.ceil(height / rowHeight);
-
-  var maxNeeded = rows * columns;
-
-  if (withPadding) {
-    maxNeeded += columns;
-  }
-
-  var potentialVisible = length - startingIndex;
-
-  return Math.max(Math.min(maxNeeded, potentialVisible), 0);
-};
-
-FixedGrid.prototype.widthAtIndex = function (/* index */) {
-  return this._elementWidth;
-};
-
-FixedGrid.prototype.heightAtIndex = function (/* index */) {
-  return this._elementHeight;
-};
-
-FixedGrid.prototype.position = function (index, width) {
-  var length = this.length();
-  if (length === 0 || index > length) {
-    rangeError(length, index);
-  }
-
-  var columns = Math.floor(width / this.widthAtIndex(index)) || 1;
-
-  var x = index % columns * this.widthAtIndex(index) | 0;
-  var y = Math.floor(index / columns) * this.heightAtIndex(index);
-
-  return new Position(x, y);
-};
-
-FixedGrid.prototype.height = function (visibleWidth) {
-  if (typeof visibleWidth !== 'number') {
-    throw new TypeError('height depends on the first argument of visibleWidth(number)');
-  }
-  var length = this.length();
-  if (length === 0) { return 0; }
-
-  var columnCount = Math.max(Math.floor(visibleWidth/this.widthAtIndex(0), 1));
-  columnCount = columnCount > 0 ? columnCount : 1;
-  var rows = Math.ceil(length/columnCount);
-  var totalHeight = rows * this.heightAtIndex(0);
-
-  return totalHeight;
-};
-
-Bin.FixedGrid = FixedGrid;
-
-Bin.Position = Position;
-Bin.Entry = Entry;
-Bin.ShelfFirst = ShelfFirst;
-
-/* global define */
-/* global module */
-/* jshint -W069 */
-if (typeof define === 'function' && define['amd']) {
-  define(function() { return Bin; });
-} else if (typeof module !== 'undefined' && module['exports']) {
-
-  module['exports'] = Bin;
-} else if (typeof this !== 'undefined') {
-  this['Bin'] = Bin;
-}
