+++ tests/index.html
@@ -1,0 +1,13 @@
+<html>
+<head>
+  <meta charset="utf-8">
+  <title>Layout Bin Packer</title>
+  <link rel="stylesheet" href="../node_modules/qunitjs/qunit/qunit.css">
+</head>
+<body>
+  <div id="qunit"></div>
+  <script src="../node_modules/qunitjs/qunit/qunit.js"></script>
+  <script src="../dist/index.js"></script>
+  <script src="./index.js"></script>
+</body>
+</html>
